+++ src/main/java/acme/features/authenticated/manager/AuthenticatedManagerCreateService.java
@@ -1,9 +1,16 @@
 package acme.features.authenticated.manager;
 
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import acme.entities.roles.Manager;
+import acme.entities.spam.Spam;
+import acme.entities.spam.Threshold;
+import acme.features.administrator.spam.AdminSpamCreateService;
+import acme.features.administrator.spam.AdminSpamRepository;
+import acme.features.administrator.threshold.ThresholdRepository;
 import acme.framework.components.Errors;
 import acme.framework.components.HttpMethod;
 import acme.framework.components.Model;
@@ -22,6 +29,15 @@ public class AuthenticatedManagerCreateService implements AbstractCreateService<
 
 	@Autowired
 	protected AuthenticatedManagerRepository repository;
+	
+	@Autowired
+	private AdminSpamRepository spamRepository;
+	
+	@Autowired
+	protected AdminSpamCreateService spamService;
+	
+	@Autowired
+	protected ThresholdRepository	thresholdRepository;
 
 	// AbstractCreateService<Authenticated, Manager> ---------------------------
 
@@ -38,6 +54,19 @@ public class AuthenticatedManagerCreateService implements AbstractCreateService<
 		assert request != null;
 		assert entity != null;
 		assert errors != null;
+		
+		final List<Spam> spam= (List<Spam>) this.spamRepository.findSpam();
+		final Threshold threshold=this.thresholdRepository.findSpamEntity(35);
+		final boolean censuraCompany = Threshold.censura(entity.getCompany(), spam, threshold.getThreshold());
+		final boolean censuraSector = Threshold.censura(entity.getSector(), spam, threshold.getThreshold());
+		
+		if(censuraCompany) {
+			errors.add("company", "this company is spam ");
+		}
+
+		if(censuraSector) {
+			errors.add("sector", "this sector is spam ");
+		}
 	}
 
 	@Override
