+++ src/test/java/acme/testing/admin/ThresholdUpdateTest.java
@@ -1,4 +1,4 @@
-package admin;
+package acme.testing.admin;
 
 import org.junit.jupiter.api.Order;
 import org.junit.jupiter.params.ParameterizedTest;
