+++ src/test/resources/manager/manager/create-negative.csv
@@ -1,4 +1,5 @@
 recordIndex,company,sector
 0,,developers
 0,ayesa,
-
+0,sex, developer
+0,ayesa, sex
